backends/LoyalGuard/LoyalGuard.Infrastructure.Test/Sqlite/Service/LGRoleServiceTest.cs

using System.Collections.Generic;
using System.Reflection;
using Xunit;
using Serilog;
using Brash.Infrastructure;
using Brash.Infrastructure.Sqlite;
using LoyalGuard.Domain.Model;
using LoyalGuard.Infrastructure.Sqlite.Repository;
using LoyalGuard.Infrastructure.Sqlite.RepositorySql;
using LoyalGuard.Infrastructure.Sqlite.Service;
using LoyalGuard.Infrastructure.Test.Sqlite.Faker;

namespace LoyalGuard.Infrastructure.Test.Sqlite.Service
{
	public class LGRoleServiceTest
	{
		public string GetDatabase(string path, MethodBase methodBase)
		{
			string dbName = $"{methodBase.ReflectedType.Name}_{methodBase.Name}";
			string databaseFile = $"{path}/{dbName}.sqlite";
			System.IO.File.Delete(databaseFile);

			return databaseFile;
		}

		public static ILogger GetLogger(string filename)
		{
			return new LoggerConfiguration()
				.MinimumLevel.Verbose()
				.WriteTo.File($"{filename}", rollingInterval: RollingInterval.Day)
				.CreateLogger();
		}

		[Fact]
		public void CreateUpdateDeleteFetch()
		{
			// file system
			var path = "/shop/randomsilo/modern-web/backends/LoyalGuard";
			var project = "LoyalGuard";
			var outputPath = $"{path}/{project}.Infrastructure.Test/TestOutput/";
			var databaseFile = GetDatabase(outputPath, MethodBase.GetCurrentMethod());
			
			// logger
			ILogger logger = GetLogger($"{outputPath}/{MethodBase.GetCurrentMethod()}.log");
			
			// database setup

			// - context
			IDatabaseContext databaseContext = new DatabaseContext(
				$"Data Source={databaseFile}" 
				, "TestDb"
				, "TestSchema"
				, $"{path}/sql/sqlite/ALL.sql"
			);
			Assert.NotNull(databaseContext);

			// - manager
			IManageDatabase databaseManager = new DatabaseManager(databaseContext);
			Assert.NotNull(databaseManager);

			// - create tables
			databaseManager.CreateDatabase();

			// - repository
			var lGRoleRepository = new LGRoleRepository(databaseManager, new LGRoleRepositorySql(), logger);
			Assert.NotNull(lGRoleRepository);

			// - service
			var lGRoleService = new LGRoleService(lGRoleRepository, logger);
			Assert.NotNull(lGRoleService);

			// faker
			BrashActionResult<LGRole> serviceResult = null;
			var lGRoleFaker = new LGRoleFaker(databaseManager, logger);
			Assert.NotNull(lGRoleFaker);

			// create
			var lGRoleCreateModel = lGRoleFaker.GetOne();
			serviceResult = lGRoleService.Create(lGRoleCreateModel);
			Assert.True(serviceResult.Status == BrashActionStatus.SUCCESS, serviceResult.Message);
			Assert.True(serviceResult.Model.LGRoleId > 0);

			// use model with id
			lGRoleCreateModel = serviceResult.Model;

			// update
			var lGRoleUpdateModel = lGRoleFaker.GetOne();
			lGRoleUpdateModel.LGRoleId = lGRoleCreateModel.LGRoleId;
			serviceResult = lGRoleService.Update(lGRoleUpdateModel);
			Assert.True(serviceResult.Status == BrashActionStatus.SUCCESS, serviceResult.Message);

			// delete
			serviceResult = lGRoleService.Delete(lGRoleCreateModel);
			Assert.True(serviceResult.Status == BrashActionStatus.SUCCESS, serviceResult.Message);

			// fetch

			// - make fakes
			var fakes = lGRoleFaker.GetMany(10);

			// - add fakes to database
			List<int?> ids = new List<int?>();
			foreach (var f in fakes)
			{
				serviceResult = lGRoleService.Create(f);

				Assert.True(serviceResult.Status == BrashActionStatus.SUCCESS, serviceResult.Message);
				Assert.True(serviceResult.Model.LGRoleId >= 0);
				ids.Add(serviceResult.Model.LGRoleId);
			}

			// - get fakes from database
			foreach(var id in ids)
			{
				var model = new LGRole()
				{
					LGRoleId = id
				};

				serviceResult = lGRoleService.Fetch(model);
				Assert.True(serviceResult.Status == BrashActionStatus.SUCCESS, serviceResult.Message);
				Assert.True(serviceResult.Model.LGRoleId >= 0);
			}
		}

	}
}